Fix process leakage.
The per-process log processing thread of each started process
was 'link'ed to the parent thread, which is itself responsible
for orchestrating the lifecycle of all processes. These 'Async's
are 'cancel'ed as processes finish or are terminated. Thus, as soon
as the first process finishes and the first such Async is canceled,
the parent thread responsible for the orchestrating (i.e. currently
the main thread) is itself killed, resulting in frequent orphan
processes as well as keiretsu exiting with "keiretsu: thread killed".

Instead of linking the threads, just log any unexpected errors
which caused termination of the log processing threads.